+++ test/babashka/uberjar_test.clj
@@ -4,12 +4,19 @@
    [clojure.string :as str]
    [clojure.test :as t :refer [deftest is testing]]))
 
+(defn count-entries [jar]
+  (with-open [jar-file (java.util.jar.JarFile. jar)]
+    (count (map #_prn
+                identity
+                (enumeration-seq
+                 (.entries jar-file))))))
+
 (deftest uberjar-test
   (testing "uberjar with --main"
     (let [tmp-file (java.io.File/createTempFile "uber" ".jar")
           path (.getPath tmp-file)]
       (.deleteOnExit tmp-file)
-      (tu/bb nil "uberjar" path "--classpath" "test-resources/babashka/uberjar/src" "-m" "my.main-main")
+      (tu/bb nil "--classpath" "test-resources/babashka/uberjar/src" "uberjar" path "-m" "my.main-main")
       (is (= "(\"1\" \"2\" \"3\" \"4\")\n"
              (tu/bb nil "--jar" path "1" "2" "3" "4")))
       (is (= "(\"1\" \"2\" \"3\" \"4\")\n"
@@ -24,7 +31,7 @@
     (let [tmp-file (java.io.File/createTempFile "uber" ".jar")
           path (.getPath tmp-file)]
       (.deleteOnExit tmp-file)
-      (tu/bb nil "uberjar" path "--classpath" "test-resources/babashka/uberjar/src")
+      (tu/bb nil "--classpath" "test-resources/babashka/uberjar/src" "uberjar" path)
       (is (str/includes? (tu/bb "(+ 1 2 3)" path) "6"))))
   (testing "use bb.edn classpath when no other --classpath"
     (tu/with-config {:paths ["test-resources/babashka/uberjar/src"]}
@@ -36,4 +43,18 @@
         ;; running
         (is (= "(\"42\")\n" (tu/bb nil "--jar" path "-m" "my.main-main" "42")))
         (is (= "(\"42\")\n" (tu/bb nil "--classpath" path "-m" "my.main-main" "42")))
-        (is (= "(\"42\")\n" (tu/bb nil path "42")))))))
+        (is (= "(\"42\")\n" (tu/bb nil path "42"))))))
+  (testing "throw on empty classpath"
+    (let [tmp-file (java.io.File/createTempFile "uber" ".jar")
+          path (.getPath tmp-file)]
+      (.deleteOnExit tmp-file)
+      (is (thrown-with-msg?
+           Exception #"classpath"
+           (tu/bb nil "uberjar" path "-m" "my.main-main")))))
+  (testing "ignore empty entries on classpath"
+    (let [tmp-file (java.io.File/createTempFile "uber" ".jar")
+          path (.getPath tmp-file)]
+      (.deleteOnExit tmp-file)
+      (tu/bb nil "--classpath" ":::" "uberjar" path "-m" "my.main-main")
+      ;; Only a manifest entry is added
+      (is (< (count-entries path) 3)))))